Move scaffold sample into separate file
